requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# ML & CV - Core
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.4.0

# Phase 1: Analysis Models
open-clip-torch>=2.24.0
timm>=0.9.12

# AWS (S3 and SQS)
boto3>=1.34.0

# HTTP Client
httpx>=0.26.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
